requirements.txt
twitchio
aiohttp
python-dotenv